Split demo app
