+++ src/demoapp/DocViewerWithInputApp.tsx
@@ -1,0 +1,30 @@
+import React, { useState } from "react";
+import DocViewer from "../index";
+import { DocViewerRenderers } from "../plugins";
+
+const DocViewerWithInputApp = () => {
+  const [selectedDocs, setSelectedDocs] = useState<File[]>([]);
+
+  return (
+    <>
+      <input
+        type="file"
+        accept=".pdf"
+        multiple
+        onChange={(el) =>
+          el.target.files?.length &&
+          setSelectedDocs(Array.from(el.target.files))
+        }
+      />
+      <DocViewer
+        documents={selectedDocs.map((file) => ({
+          uri: window.URL.createObjectURL(file),
+          fileName: file.name,
+        }))}
+        pluginRenderers={DocViewerRenderers}
+      />
+    </>
+  );
+};
+
+export default DocViewerWithInputApp;
