+++ src/demoapp/index.tsx
@@ -1,53 +1,11 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import DocViewer from "..";
-import { DocViewerRenderers } from "../plugins";
 
-const App = () => {
-  const docs = [
-    { uri: require("./exampleFiles/pdf-file.pdf") },
-    { uri: require("./exampleFiles/png-image.png") },
-    { uri: require("./exampleFiles/csv-file.csv") },
-  ];
-
-  return (
-    <DocViewer
-      documents={docs}
-      initialActiveDocument={docs[1]}
-      pluginRenderers={DocViewerRenderers}
-      config={{
-        noRenderer: {
-          overrideComponent: ({ document, fileName }) => {
-            const fileText = fileName || document?.fileType || "";
-            console.log(document);
-            if (fileText) {
-              return <div>no renderer for {fileText}</div>;
-            }
-            return <div>no renderer</div>;
-          },
-        },
-        loadingRenderer: {
-          overrideComponent: ({ document, fileName }) => {
-            const fileText = fileName || document?.fileType || "";
-            if (fileText) {
-              return <div>loading ({fileText})</div>;
-            }
-            return <div>loading</div>;
-          },
-        },
-        csvDelimiter: ",",
-        pdfZoom: {
-          defaultZoom: 1.1,
-          zoomJump: 0.2,
-        },
-      }}
-      language="pl"
-    />
-  );
-};
+// import DocViewerApp from "./DocViewerApp";
+import DocViewerWithInputApp from "./DocViewerWithInputApp";
 
 const rootElement = document.getElementById("root");
 if (rootElement) {
   const root = createRoot(rootElement);
-  root.render(<App />);
+  root.render(<DocViewerWithInputApp />);
 }
